requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
Django==3.1
Flask==1.1.2
idna==2.10
image==1.5.32
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.0
networkx==2.0
numpy==1.15.0
opencv-python==4.3.0.38
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
requests==2.24.0
scikit-image==0.13.1
scipy==1.5.2
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
Werkzeug==0.15.5